Removed file from wrong directory. Added overide css for the presentation of the transparent background
