+++ includes/blocks/happy_point.twig
@@ -1,16 +1,35 @@
 {% block happy_point %}
     {% if ( fields ) %}
-
+	<style>
+		@media (min-width: 992px) {
+			.eager-wrap {
+				background: url("{{ fields['background_src'][0] }}") center center no-repeat;
+			}
+			.eager-wrap:before
+			{
+				opacity: {{ fields['opacity']}};
+			}
+		}
+		@media (min-width: 1200px) {
+			.eager-wrap {
+				background: url("{{ fields['background_src'][0] }}") center center no-repeat;
+			}
+			.eager-wrap:before
+			{
+				opacity: {{ fields['opacity']}};
+			}
+		}
+	</style>
 		<div class="eager-wrap">
 			<div class="container">
 				{% if ( fields['boxout_title'] ) %}
 					<div class="eager-for-more hidden-md-down">
-						<h2>{{ fields['boxout_title']|e('html') }}</h2>
+						<h2>{{ fields['boxout_title']}}</h2>
 						{% if ( fields['boxout_descr'] ) %}
-							<h5>{{ fields['boxout_descr']|e('html') }}</h5>
+							<h5>{{ fields['boxout_descr'] }}</h5>
 						{% endif %}
 						{% if (( fields['boxout_link_url'] ) and (fields['boxout_link_text'])) %}
-							<a href="{{ fields['boxout_link_url']|e('html_attr') }}" class="btn btn-primary">{{ fields['boxout_link_text']|e('html') }}</a>
+							<a href="{{ fields['boxout_link_url']}}" class="btn btn-primary">{{ fields['boxout_link_text']}}</a>
 						{% endif %}
 					</div>
 				{% endif %}
@@ -19,16 +38,16 @@
 			<div class="eager-wrap-mobile hidden-lg-up">
 				<div class="container">
 					{% if ( fields['background'] ) %}
-						{{ fields['background'] }}
+						{{ fields['background_html']|raw }}
 					{% endif %}
 					{% if ( fields['boxout_title'] ) %}
-						<h3>{{ fields['boxout_title']|e('html') }}</h3>
+						<h3>{{ fields['boxout_title']}}</h3>
 					{% endif %}
 					{% if ( fields['boxout_descr'] ) %}
-						<h4>{{ fields['boxout_descr']|e('html') }}</h4>
+						<h4>{{ fields['boxout_descr']}}</h4>
 					{% endif %}
 					{% if (( fields['boxout_link_url'] ) and (fields['boxout_link_text'])) %}
-						<a href="{{ fields['boxout_link_url']|e('html_attr') }}" class="btn btn-primary">{{ fields['boxout_link_text']|e('html') }}</a>
+						<a href="{{ fields['boxout_link_url']}}" class="btn btn-primary">{{ fields['boxout_link_text']}}</a>
 					{% endif %}
 				</div>
 			</div>
